src/app/dashboards/azure.json: Consolidate timestamp and filter rows
Newer versions of Kibana have Query and Filtering rows built into the
title bar, so we don't need to code them into our dashboard directly.
This commit drops our local Query bar and shifts the Filtering blurb
up to the same row as the timestamp blurb.  I've also used the 'title'
attribute instead of coding the title in Markdown, and updated the
filtering text to match the new title-bar implementation.